job、order表新增字段，修改了mapper文件；修改部分主键自增；新增部分订单子系统api。

+++ src/main/java/com/jbgz/pancakejob/entity/Order.java
@@ -55,12 +55,12 @@ public class Order implements Serializable {
     /**
      * 
      */
-    private BigDecimal jobhunterScore;
+    private Integer jobhunterScore;
 
     /**
      * 
      */
-    private BigDecimal recuriterScore;
+    private Integer recuriterScore;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
